show current username/unable to delete urself

+++ src/main/java/com/lit/hw/controller/MainController.java
@@ -1,7 +1,15 @@
 package com.lit.hw.controller;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,11 +25,14 @@ public class MainController {
 
 	@Autowired
 	UserService userService;
+	static String message;
 
-	///////////////////////////////////////////////////////////// page
-	
+	// /////////////////////////////////////////////////////////// page
 	@RequestMapping(value = "/", method = RequestMethod.GET)
 	public String homePage(ModelMap model) {
+		model.addAttribute("message", message);
+		message = "";
+		model.addAttribute("currentUser", userService.show(getCurrentUsername()));
 		model.addAttribute("listUser", userService.list());
 		System.out.println("home page");
 		return "home";
@@ -29,6 +40,7 @@ public class MainController {
 
 	@RequestMapping(value = "/add", method = RequestMethod.GET)
 	public String addPage(ModelMap model) {
+		model.addAttribute("currentUser", userService.show(getCurrentUsername()));
 		model.addAttribute("page", "Add");
 		System.out.println("add page");
 		return "_form";
@@ -36,6 +48,7 @@ public class MainController {
 
 	@RequestMapping(value = "/update/{id}", method = RequestMethod.GET)
 	public String updatePage(ModelMap model, @PathVariable int id) {
+		model.addAttribute("currentUser", userService.show(getCurrentUsername()));
 		model.addAttribute("page", "Update");
 		model.addAttribute("usrObj", userService.show(id));
 		System.out.println("update page");
@@ -44,6 +57,7 @@ public class MainController {
 
 	@RequestMapping(value = "/show/{id}", method = RequestMethod.GET)
 	public String showPage(ModelMap model, @PathVariable int id) {
+		model.addAttribute("currentUser", userService.show(getCurrentUsername()));
 		model.addAttribute("page", "Show");
 		model.addAttribute("readonly", "readonly");
 		model.addAttribute("disabled", "disabled");
@@ -52,7 +66,7 @@ public class MainController {
 		return "_form";
 	}
 
-	///////////////////////////////////////////////////////////// action
+	// /////////////////////////////////////////////////////////// action
 	@RequestMapping(value = "/add", method = RequestMethod.POST)
 	public String add(@ModelAttribute(value = "usr") User usrObj) {
 		System.out.println("add action " + usrObj.getUsername());
@@ -68,20 +82,67 @@ public class MainController {
 	}
 
 	@RequestMapping(value = "/delete/{id}", method = RequestMethod.GET)
-	public String delete(@PathVariable int id) {
-		System.out.println("delete action" + id);
-		userService.delete(id);
+	public String delete(ModelMap model, @PathVariable int id) {
+		if (userService.show(getCurrentUsername()).getId() == id) {
+			message = "You cannot delete yourself.";
+		} else {
+			System.out.println("delete action" + id);
+			userService.delete(id);
+		}
 		return "redirect:/";
 	}
 
 	@RequestMapping(value = "/search", method = RequestMethod.POST)
-	public String search(ModelMap model, 
-			@RequestParam(value = "keyword", required = true) String keyword,
+	public String search(ModelMap model, @RequestParam(value = "keyword", required = true) String keyword,
 			@RequestParam(value = "type", required = true) String type) {
 		System.out.println("search action " + keyword + " " + type);
-		if(!keyword.equals("")){ 
+		if (!keyword.equals("")) {
 			model.addAttribute("listUser", userService.search(keyword, type));
 			return "home";
-		}else return "redirect:/";
+		} else
+			return "redirect:/";
 	}
+
+	// ////////////////////////////////////////////////security page
+	@RequestMapping(value = "/403")
+	public String accessDenied(ModelMap m) {
+		System.out.println("access denied 403");
+		return "403";
+	}
+
+	@RequestMapping(value = "/login")
+	public String loginPage(Model model) {
+		if (isAuthenticated())
+			return "redirect:/";
+		System.out.println("login page");
+		return "login";
+	}
+
+	@RequestMapping(value = "/logout")
+	public String logoutAction(HttpServletRequest request, HttpServletResponse response) {
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		if (auth != null) {
+			new SecurityContextLogoutHandler().logout(request, response, auth);
+			System.out.println("success logout");
+		}
+		return "redirect:/";
+	}
+
+	// ////////////////////////////////////////////////other function
+	public boolean isAuthenticated() {
+		return !(SecurityContextHolder.getContext().getAuthentication() instanceof AnonymousAuthenticationToken);
+	}
+
+	private String getCurrentUsername() {
+		return SecurityContextHolder.getContext().getAuthentication().getName();
+	}
+	/*
+	 * private boolean isLogin() { return
+	 * SecurityContextHolder.getContext().getAuthentication().getName() ==
+	 * "anonymousUser" ? false : true; }
+	 * 
+	 * private String getRole() { return
+	 * SecurityContextHolder.getContext().getAuthentication
+	 * ().getAuthorities().toString(); }
+	 */
 }
